Move target contexts into interaction
