+++ engine/src/main/kotlin/world/gregs/voidps/engine/client/ui/interact/ItemOnPlayer.kt
@@ -1,8 +1,8 @@
 package world.gregs.voidps.engine.client.ui.interact
 
 import world.gregs.voidps.engine.entity.character.Character
-import world.gregs.voidps.engine.entity.character.TargetPlayerContext
 import world.gregs.voidps.engine.entity.character.mode.interact.Interaction
+import world.gregs.voidps.engine.entity.character.mode.interact.TargetPlayerContext
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.item.Item
