--- engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/TargetNPCContext.kt
@@ -1,7 +1,0 @@
-package world.gregs.voidps.engine.entity.character
-
-import world.gregs.voidps.engine.entity.character.npc.NPC
-
-interface TargetNPCContext : CharacterContext {
-    val target: NPC
-}
