+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/TargetFloorItemContext.kt
@@ -1,5 +1,6 @@
-package world.gregs.voidps.engine.entity.character
+package world.gregs.voidps.engine.entity.character.mode.interact
 
+import world.gregs.voidps.engine.entity.character.CharacterContext
 import world.gregs.voidps.engine.entity.item.floor.FloorItem
 
 interface TargetFloorItemContext : CharacterContext {
