+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/mode/interact/TargetNPCContext.kt
@@ -1,0 +1,8 @@
+package world.gregs.voidps.engine.entity.character.mode.interact
+
+import world.gregs.voidps.engine.entity.character.CharacterContext
+import world.gregs.voidps.engine.entity.character.npc.NPC
+
+interface TargetNPCContext : CharacterContext {
+    val target: NPC
+}
