+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/player/PlayerOption.kt
@@ -1,8 +1,8 @@
 package world.gregs.voidps.engine.entity.character.player
 
 import world.gregs.voidps.engine.entity.character.Character
-import world.gregs.voidps.engine.entity.character.TargetPlayerContext
 import world.gregs.voidps.engine.entity.character.mode.interact.Interaction
+import world.gregs.voidps.engine.entity.character.mode.interact.TargetPlayerContext
 
 data class PlayerOption(
     override val character: Character,
