+++ game/src/main/kotlin/world/gregs/voidps/world/map/al_kharid/duel_arena/Nurses.kt
@@ -1,8 +1,8 @@
 package world.gregs.voidps.world.map.al_kharid.duel_arena
 
 import world.gregs.voidps.engine.client.message
-import world.gregs.voidps.engine.entity.character.TargetNPCContext
 import world.gregs.voidps.engine.entity.character.face
+import world.gregs.voidps.engine.entity.character.mode.interact.TargetNPCContext
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
 import world.gregs.voidps.engine.entity.character.setAnimation
 import world.gregs.voidps.world.interact.dialogue.Cheerful
